Title: 자연수 n개의 합, Time: 129ms, Memory: 10MB, Status: Passed - Codetree

+++ 250604/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -16,7 +16,7 @@ public class Main {
 
             long mid = left + (right - left) / 2;
 
-            if (mid * (mid + 1) / 2 <= s) {
+            if ((double) mid * (mid + 1) / 2 <= s) {
                 left = mid + 1;
                 res = Math.max(res, mid);
